+++ services/authentication-service/internal/staff/errors.go
@@ -6,6 +6,4 @@ import "errors"
 var (
 	// ErrStaffAlreadyExists indicates that a staff with the same identifying details already exists in the system.
 	ErrStaffAlreadyExists = errors.New("staff already exists")
-	// ErrInvalidCredentials indicates that the provided credentials are invalid during authentication processes.
-	ErrInvalidCredentials = errors.New("invalid credentials")
 )
